backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.1.1
langchain==0.3.27
langchain-google-genai==2.0.8
langchain-pinecone==0.2.12
langchain-huggingface==0.3.1
pinecone-client==2.2.4
sentence-transformers==3.3.1
pydantic==2.12.2
pandas==2.2.3